requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
#django-distill==2.8
django-ultracache==2.1.1
idna==3.2
pika==0.13.1
python-memcached==1.59
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
uWSGI==2.0.19.1
-e git+ssh://git@github.com/hedleyroos/django-distill.git@master#egg=django_distill
